Adding metal support

+++ cocos2d-ui/CCBReader/CCBReader.m
@@ -167,7 +167,7 @@ static inline void alignBits(CCBReader *self)
 }
 
 
-static inline unsigned int readVariableLengthIntFromArray(const uint8_t* buffer, uint32_t * value) {
+static inline ptrdiff_t readVariableLengthIntFromArray(const uint8_t* buffer, uint32_t * value) {
     const uint8_t* ptr = buffer;
     uint32_t b;
     uint32_t result;
